实现功能可用

+++ nvm.sh
@@ -14,44 +14,54 @@ exit_if_err () {
 # @func - 需要权限时检查如果不是 root 运行就退出
 check_permission () {
     [ "root" = "$USER" ]
-    exit_if_err '非 root请使用 sudo 运行'
+    exit_if_err '非 root 请使用 sudo 运行'
 }
 
 # 定义变量
-node_link_path="/usr/local/nodejs"
-node_bin_path="$node_link_path/bin"
-node_versions_dir="/opt/node-versions"
-npm_global_dir="$node_versions_dir/npm_global_modules"
-npm="$node_bin_path/npm"
-add_node_path="/etc/profile.d/add_node_path.sh"
+node_versions_dir="/opt/nvm/node-versions"
+npm_global_dir="/opt/nvm/npm_global"
 
+node_current_path="/opt/nvm/node-current"
+node="$node_current_path/bin/node"
+
+nvm_conf_profile="/etc/profile.d/nvm-conf-profile.sh"
+
+# @func
+# 检查是否已经进行过初始化，满足条件：
+#
+# 1. $node_current_path/bin/node 可运行
+# 2. ① 指向的 node 是通过 nvm 安装的
+# 3. /usr/local/bin/[node, npm, npx] 指向 $node_current_path/bin/[x]
+# 4. npm_grobal_dir 文件夹存在
+# 5. nvm_conf_profile 存在
+#
 check_init () {
-    [ -L "$node_link_path" ] &&\
-    [ -r "$add_node_path" ] &&\
-    [ -x $npm -a -x $node_bin_path/node ] &&\
-    [ -d "$npm_global_dir" ] &&\
-    grep -q "prefix=$npm_global_dir" ~/.npmrc || {
-        echo "未初始化，先运行 nvm init"
-        return 1
-    }
+    [ -x $node ] && {
+        for x in node npm npx
+        do
+            [ /usr/local/bin/$x -ef $node_current_path/bin/$x ] || return 1
+        done
+    } &&\
+    [ -d $npm_global_dir ] &&\
+    [ -r $nvm_conf_profile ]
 }
 
 list () {
-    [ -x $node_bin_path/node ] && v_current=`$node_bin_path/node --version`
+    [ -x $node ] && v_current=`$node --version`
 	versions=`ls $node_versions_dir | grep -oP 'v(\d+\.?)+'`
-	for version in $versions;
+	for version in $versions
 	do
 		[ "$version" = "$v_current" ] && echo "* $version" || echo "  $version"
 	done
 }
 
 check_version_input () {
-    [ `echo $1 | grep -P '^v(\d+\.){2}\d+$' | wc -l` -eq 1 ]
+    [ `echo "$1" | grep -P '^v(\d+\.){2}\d+$' | wc -l` -eq 1 ]
     exit_if_err '版本号不正确 匹配 v主版本号.次版本号.修订号'
 }
 
 usage () {
-    echo "usage: nvm [ls | use | init | install] [version]"
+    echo "usage: nvm [ls | use | install | remove] [version]"
 }
 
 use () {
@@ -61,48 +71,56 @@ use () {
     # 检查输入版本号是否正确
     check_version_input $1
 
-    v_use=`list | grep $1`
+    v_use=`echo "\`list\`" | grep $1`
 
     # 没有找到版本号则退出
     [ -n "$v_use" ]
     exit_if_err '未找到该版本号'
 
     # 找到的版本号正在使用中则退出
-    [ -z "`echo $v_use | grep -P "^\*"`" ]
-    exit_if_err "正在使用 $v_use"
+    echo "$v_use" | grep -qP "^\*" && {
+        echo "正在使用 $v_use"
+        exit
+    }
 
     # 可以切换的版本号
     # 删除原有软连接并重新创建
-    [ -L "$node_link_path" ] && rm $node_link_path
+    [ -L "$node_current_path" ] && rm $node_current_path
     v_dir_name=`ls $node_versions_dir | grep $v_use`
-    ln -s $node_versions_dir/$v_dir_name $node_link_path
+    ln -s $node_versions_dir/$v_dir_name $node_current_path
     exit_if_err '重新创建软连接失败'
 }
 
+# @func
+# 在第一次通过 nvm 安装 node 时进行，
+# 使 check_init 检查可以通过
+#
 init () {
     # 检查权限
     check_permission
 
-    # 创建 npm 全局包文件夹
-    [ -d "$npm_global_dir" ] || {
+    # 1. $node_current_path/bin/node 可运行
+    # 2. ① 指向的 node 是通过 nvm 安装的
+    # $node_current_path 链接由 use 方法负责创建
+    [ -x $node ]
+    exit_if_err '未安装 node 版本'
+
+    # 3. /usr/local/bin/[node, npm, npx] 指向 $node_current_path/bin/[x]
+    for x in node npm npx
+    do
+        ln -s $node_current_path/bin/$x /usr/local/bin/$x
+        exit_if_err "创建 /usr/local/bin/$x 软连接失败"
+    done
+
+    # 4. npm_grobal_dir 文件夹存在
+    [ -d $npm_global_dir ] || {
         mkdir -p $npm_global_dir
         exit_if_err "创建 $npm_global_dir 失败"
     }
-    
-    # 如果当前已经安装 node 版本
-    # 添加 node, npm 路径到 PATH
-    # node_path=`which node | awk -F bin/node '{print $1}'`
-    [ -x $npm ] && {
-        PATH="$PATH:$node_bin_path"
-        $npm config set prefix $npm_global_dir && {
-            [ -r "$add_node_path" ] ||\
-            echo "PATH=$npm_global_dir/bin:$node_bin_path:\$PATH" >> $add_node_path &&\
-            . $add_node_path
-            exit_if_err  "创建并加载 $add_node_path 失败"
-        }
-    }
-    
-    exit_if_err "未安装 node.js, 先 nvm install {version}"
+
+    # 5. nvm_conf_profile 存在
+    [ -r $nvm_conf_profile ] || cp ./nvm-conf-profile.sh /etc/profile.d/
+    exit_if_err  "创建 $nvm_conf_profile 失败"
 }
 
 install () {
@@ -111,6 +129,13 @@ install () {
 
     # 检查输入版本号是否正确
     check_version_input $1
+
+    # 不能重复安装同一版本
+    [ -d $node_versions_dir ] &&\
+    echo "`list`" | grep -q $1 && {
+        use $1
+        exit
+    }
     
     # 机器表示，不同机器需要不同的安装包
     m_flag=`uname -m`
@@ -122,14 +147,16 @@ install () {
     exit_if_err "下载 $url 失败"
 
     # 解压放到多版本文件夹
+    [ -d $node_versions_dir ] || mkdir -p $node_versions_dir
+
     tar -xJf /tmp/node-$1.tar.xz -C $node_versions_dir || {
-        RET=$0
+        RET=$?
 
         echo -e "\n安装失败，下载的 node 安装包不完整，请重试。"
         
         # 清理垃圾
         bad_dir_name=`ls $node_versions_dir | grep $1`
-        rm -r $node_versions_dir/$bad_dir_name
+        [ -n "$bad_dir_name" ] && rm -r $node_versions_dir/$bad_dir_name
         rm /tmp/node-$1.tar.xz
 
         exit $RET
@@ -137,12 +164,11 @@ install () {
 
     # 解压完了调用 use 使可以使用
     use $1
-    exit_if_err "use 版本 $1 失败"
 
     # 第一次安装版本时需要 init 并重启终端
     check_init || {
-        init &&\
-        echo "node $1 安装成功，这是你的第一次安装 node，需要重新启动终端才能生效。"
+        init && . ./nvm-conf-profile.sh
+        echo "这是你的第一次安装 node，请重新启动终端，否则全局 npm 包将缺少全局路径。"
     }
 }
 
@@ -153,7 +179,7 @@ remove () {
     # 检查输入版本号是否正确
     check_version_input $1
 
-    [ -x $node_bin_path/node ] && v_current=`$node_bin_path/node --version`
+    [ -x $node ] && v_current=`$node --version`
     exit_if_err '未安装 node 版本'
 
     [ "$1" != "$v_current" ]
@@ -163,16 +189,11 @@ remove () {
 }
 
 case "$1" in
-init)
-    init
-    ;;
 ls | list)
-    check_init &&\
-    list
+    check_init && list || echo '未安装 node 版本'
     ;;
 use)
-    check_init &&\
-    use $2
+    check_init && use $2
     ;;
 i | install)
     install $2
